% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{create_factors}
\alias{create_factors}
\title{Transform all character strings to factors or update meta-data from existing
factors}
\usage{
create_factors(ds)
}
\arguments{
\item{ds}{Your data-set as data.frame.}
}
\value{
A data frame with factors if the input variables where either the
    class "character" or "factor".
}
\description{
This function converts ALL character-strings variables in a
    data frame into factorial variables. In addition it updates the meta-
    data from already existing factorial variables.
}
\examples{

# no example yet

}
