% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{A wrapper around the read.csv function allowing for creating factors on the fly}
\usage{
read_data(
  pathname,
  sep = ",",
  dec = ".",
  create_factors = T,
  show_structure = F
)
}
\arguments{
\item{pathname}{The path to your csv file.}

\item{sep}{The column separator, default is `,`.}

\item{dec}{the decimal separator, default is `.`.}

\item{create_factors}{If `TRUE`, all columns will be returned as factorial, default is `TRUE`.}

\item{show_structure}{Shows the structure of the final data via `str()`.}

\item{show_head}{Shows the head of the final data via `head()`.}
}
\value{
A data frame with factors if the input data was non-numbers (integers or floats).
}
\description{
A wrapper around the read.csv function allowing for creating factors on the fly
}
\examples{

# no example yet

}
