% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_random.R
\name{select_random}
\alias{select_random}
\title{Automatically select the random structure using "lme"}
\usage{
select_random(
  dat,
  fixed_structure,
  random_variables,
  random_slopes = F,
  keep = F,
  sort_by = "BIC",
  compute_gls = T,
  MC = T,
  noC = parallel::detectCores()
)
}
\arguments{
\item{dat}{The data to use. Should be a data frame.}

\item{fixed_structure}{The fixed structure of the statistical model, e.g. `z ~ x + y`. Must have a match in the headers of `dat`.}

\item{random_variables}{The variables that should be used, e.g. `c("d", "e")`. Must have a match in the headers of `dat`.}

\item{random_slopes}{The fixed independent variables that should be affected by `random_variables`, e.g.: `c("x", "y")`. The default is to the logical `F`. Must have a match in the headers of `dat`. Default}

\item{keep}{Should **any** variable from `random_variables` or `random_slopes` not be excluded? Default to the logical `F`, else a vector of character strings, e.g. c("d").}

\item{sort_by}{Should the data be sorted by `AIC` or `BIC`? Default is to BIC.}

\item{compute_gls}{Should a model without any random structure be included? Default to `T`.}

\item{MC}{Should multiple cores be used? The function runs per default (`MC = T`) in parallel if enough random variables are added.}

\item{noC}{The number of cores that should be used. The default setting uses all possible cores/threads: `parallel::detectCores()`. For a smaller number of random variables, a lower core count might increase the speed of the function.}
}
\value{
A data frame including the (1) random structure, (2) the AIC values, (3) the BIC values, (4) the dAIC values, (5) the dBIC values. The default sorting is by BIC.
}
\description{
Automatically select the random structure using "lme"
}
\examples{

ds <- data.frame(distance = Orthodont$distance,
                 age = Orthodont$age,
                 sex = Orthodont$Sex,
                 subject = Orthodont$Subject)

select_random(dat = ds,
              fixed_structure = c("distance ~ age * sex"),
              random_variables = c("subject"),
              random_slopes = c("age"))


select_random(dat = ds,
              fixed_structure = c("distance ~ age"),
              random_variables = c("subject", "sex"),
              random_slopes = c("age"))
}
